Extract Deposit extra collateral modal
