+++ web/components/manage-vault.tsx
@@ -3,6 +3,7 @@ import { Modal } from '@components/ui/modal';
 import { Tooltip } from '@blockstack/ui';
 import { InformationCircleIcon } from '@heroicons/react/solid';
 import { Container } from './home';
+import { VaultDepositModal } from '@components/vault-deposit-modal';
 import { stacksNetwork as network } from '@common/utils';
 import { useSTXAddress } from '@common/use-stx-address';
 import { useConnect } from '@stacks/connect-react';
@@ -16,7 +17,6 @@ import {
   makeStandardFungiblePostCondition,
   createAssetInfo,
   FungibleConditionCode,
-  makeStandardSTXPostCondition,
   falseCV,
 } from '@stacks/transactions';
 import { AppContext, CollateralTypeProps } from '@common/context';
@@ -30,7 +30,6 @@ import {
 } from '@common/vault-utils';
 import { Redirect } from 'react-router-dom';
 import { resolveReserveName, tokenTraits } from '@common/vault-utils';
-import BN from 'bn.js';
 import { tokenList } from '@components/token-swap-list';
 import { InputAmount } from './input-amount';
 import { getRPCClient } from '@common/utils';
@@ -324,65 +323,7 @@ export const ManageVault = ({ match }) => {
     }
   }, [state.currentTxStatus]);
 
-  const addDeposit = async () => {
-    if (!extraCollateralDeposit) {
-      return;
-    }
-    const token = tokenTraits[vault['collateralToken'].toLowerCase()]['name'];
-
-    let postConditions: any[] = [];
-    if (vault['collateralToken'].toLowerCase() === 'stx') {
-      postConditions = [
-        makeStandardSTXPostCondition(
-          senderAddress || '',
-          FungibleConditionCode.Equal,
-          new BN(parseFloat(extraCollateralDeposit) * 1000000)
-        ),
-      ];
-    } else {
-      // postConditions = [
-      //   makeStandardFungiblePostCondition(
-      //     senderAddress || '',
-      //     FungibleConditionCode.Equal,
-      //     new BN(parseFloat(extraCollateralDeposit) * 1000000),
-      //     createAssetInfo(
-      //       "CONTRACT_ADDRESS",
-      //       token,
-      //       vault['collateralToken'].toUpperCase()
-      //     )
-      //   )
-      // ];
-    }
-
-    await doContractCall({
-      network,
-      contractAddress,
-      stxAddress: senderAddress,
-      contractName: 'arkadiko-freddie-v1-1',
-      functionName: 'deposit',
-      functionArgs: [
-        uintCV(match.params.id),
-        uintCV(parseFloat(extraCollateralDeposit) * 1000000),
-        contractPrincipalCV(process.env.REACT_APP_CONTRACT_ADDRESS || '', reserveName),
-        contractPrincipalCV(process.env.REACT_APP_CONTRACT_ADDRESS || '', token),
-        contractPrincipalCV(
-          process.env.REACT_APP_CONTRACT_ADDRESS || '',
-          'arkadiko-collateral-types-v1-1'
-        ),
-      ],
-      postConditions,
-      onFinish: data => {
-        console.log('finished deposit!', data);
-        setState(prevState => ({
-          ...prevState,
-          currentTxId: data.txId,
-          currentTxStatus: 'pending',
-        }));
-        setShowDepositModal(false);
-      },
-      anchorMode: AnchorMode.Any,
-    });
-  };
+  
 
   const liquidationPrice = () => {
     if (vault) {
@@ -417,9 +358,7 @@ export const ManageVault = ({ match }) => {
 
   const onInputChange = (event: { target: { value: any; name: any } }) => {
     const value = event.target.value;
-    if (event.target.name === 'depositCollateral') {
-      setExtraCollateralDeposit(value);
-    } else if (event.target.name === 'mintDebt') {
+    if (event.target.name === 'mintDebt') {
       setUsdToMint(value);
     } else if (event.target.name === 'burnDebt') {
       setUsdToBurn(value);
@@ -594,10 +533,6 @@ export const ManageVault = ({ match }) => {
     });
   };
 
-  const depositMaxAmount = () => {
-    setExtraCollateralDeposit(state.balance['stx'] / 1000000 - 1);
-  };
-
   const mintMaxAmount = () => {
     setUsdToMint(
       availableCoinsToMint(
@@ -684,51 +619,12 @@ export const ManageVault = ({ match }) => {
     <Container>
       {auctionEnded && <Redirect to="/vaults" />}
 
-      <Modal
-        open={showDepositModal}
-        title="Deposit Extra Collateral"
-        icon={<img className="w-10 h-10 rounded-full" src={tokenList[2].logo} alt="" />}
-        closeModal={() => setShowDepositModal(false)}
-        buttonText="Add deposit"
-        buttonAction={() => addDeposit()}
-        initialFocus={inputRef}
-      >
-        <p className="text-sm text-center text-gray-500">
-          Choose how much extra collateral you want to post. You have a balance of{' '}
-          {state.balance[vault?.collateralToken.toLowerCase()] / decimals}{' '}
-          {vault?.collateralToken.toUpperCase()}.
-        </p>
-        <p className="text-sm text-center text-gray-500">
-          We will automatically harvest any DIKO you are eligible for when depositing.
-        </p>
-
-        <div className="mt-4">
-          <Alert>
-            <p>
-              When depositing in a vault that is already stacking, keep in mind that your extra collateral will be <span className="font-semibold">locked but not stacked</span>. You won't be able to stack these STX until the cooldown cycle!
-            </p>
-          </Alert>
-        </div>
-
-        <div className="mt-6">
-          <InputAmount
-            balance={(
-              state.balance[vault?.collateralToken.toLowerCase()] / decimals
-            ).toLocaleString(undefined, {
-              minimumFractionDigits: 2,
-              maximumFractionDigits: 6,
-            })}
-            token={vault?.collateralToken.toUpperCase()}
-            inputName="depositCollateral"
-            inputId="depositExtraStxAmount"
-            inputValue={extraCollateralDeposit}
-            inputLabel="Deposit Extra Collateral"
-            onInputChange={onInputChange}
-            onClickMax={depositMaxAmount}
-            ref={inputRef}
-          />
-        </div>
-      </Modal>
+      {/* TODO Add Show DepositModal Call */}
+      <VaultDepositModal
+        showDepositModal={showDepositModal}
+        setShowDepositModal={setShowDepositModal}
+        match={match}
+      />
 
       <Modal
         open={showWithdrawModal}
